fix: fixes collisions between `$config` properties of OAuth and parent HTTP client
2.1.0 made changes to internal protected members, refactoring those that were prefixed with an underscore to remove the underscore.
Unfortunately, one of these, `Laminas\OAuth\Client::$config`, is also defined in the parent `Laminas\Http\Client`, and the two were of different types, leading to type errors (see laminas/laminas-twitter#6).

I've audited to determine what collisions exist, and this was the only one found.
As such, I've renamed the property, as well as all references to it internally, to avoid the collision.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>